Move component.ts to new testing folder
