+++ packages/nimble-components/src/select/tests/select.spec.ts
@@ -10,7 +10,7 @@ import { FilterMode, SelectFilterInputEventDetail } from '../types';
 import {
     createEventListener,
     waitAnimationFrame
-} from '../../utilities/tests/component';
+} from '../../utilities/testing/component';
 import { filterSearchLabel } from '../../label-provider/core/label-tokens';
 import { ListOptionGroup } from '../../list-option-group';
 import type { Button } from '../../button';
