+++ packages/nimble-components/src/table/tests/table-sorting.spec.ts
@@ -9,7 +9,7 @@ import {
     TableRecord
 } from '../types';
 import { SortedColumn, TablePageObject } from '../testing/table.pageobject';
-import { createEventListener } from '../../utilities/tests/component';
+import { createEventListener } from '../../utilities/testing/component';
 
 interface SimpleTableRecord extends TableRecord {
     id: string;
